[REF] Move gourping to utils module
Moved the grouping function to a new file/module utils.py.
So it can be used as a general helper function by other plugins and parsers.
