+++ src/invoice2data/extract/parsers/regex.py
@@ -19,6 +19,8 @@ from typing import Dict
 from typing import List
 from typing import Optional
 
+from ..utils import _apply_grouping
+
 
 logger = logging.getLogger(__name__)
 
@@ -104,30 +106,6 @@ def _apply_type_coercion(
     if "type" in settings:
         for k, v in enumerate(result):
             result[k] = template.coerce_type(v, settings["type"])
-    return result
-
-
-def _apply_grouping(settings: Dict[str, Any], result: Any) -> Optional[Any]:
-    """Apply grouping to the extracted values."""
-    if "group" in settings:
-        result = list(filter(None, result))
-        if result:
-            if settings["group"] == "sum":
-                result = sum(result)
-            elif settings["group"] == "min":
-                result = min(result)
-            elif settings["group"] == "max":
-                result = max(result)
-            elif settings["group"] == "first":
-                result = result[0]
-            elif settings["group"] == "last":
-                result = result[-1]
-            elif settings["group"] == "join":
-                joined = " ".join(str(v) for v in result) if result else ""
-                result = [joined]
-            else:
-                logger.warning("Unsupported grouping method: %s", settings["group"])
-                return None
     return result
